Bug 4620, Bug 4487 - updated navigation/menu links to all be <a4j:htmlCommandLink>.
For some reason the behavior of the tables in the "Reports" and "Navigation" sections were different when the back button was pressed. After looking into to it for sometime, I determined that it was caused by the way we were navigating to the pages. Using any type of JSF component that navigates you to the a page with a table through an action will cause one behavior compared to using an <a> tag or <h:outputLink> tag (which generates an <a> tag).