Reviewed By Jason Wimmer: Adding new JBehave classes to handle page objects in addition to the traditional Steps files.
The general form for all steps is "Given|When|Then|And I actionToPerform [actionParameters (varName|"") ] on an elementName of elementType [as|is (varName|"")][and ActionToPerform  [actionParameters (varName|"") ]]


See : https://sites.google.com/a/inthinc.com/portal/documentation/automation-standards

Page objects were changed to have each of them return a valid URL that can be compared to the webPage that matches the object.  If the URL contained a database ID, then the number and preceding slash are removed.